requirements.txt
torf>=3.1.0
tmdbv3api>=1.9.0
qbittorrent-api>=2023.11.0
pyyaml>=6.0.1
click>=8.1.7
requests>=2.31.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
